수정 : 형변환 관련 경고 및 모든 경고 제거
.

+++ MiniGameServer/Player.cpp
@@ -273,12 +273,12 @@ bool Player::checkCollide(const float* oppoPosVec, Packet::PacketID packetId)
 	}
 	case Packet::PacketID::PPCOLLIDERESULT:
 	{
-		float sqaureDist = 0.0f;
+		double sqaureDist = 0.0f;
 		for (int i = 0; i < 2; ++i)
 		{
 			sqaureDist += pow(m_position[i] - oppoPosVec[i], 2);
 		}
-		float testValue = pow((ServerProtocol::PLAYER_COLLIDER_RADIUS * 2.0f), 2);
+		double testValue = pow((ServerProtocol::PLAYER_COLLIDER_RADIUS * 2.0f), 2);
 		return testValue >= sqaureDist;
 	}
 	default:
